+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/SnapshotChangeAuditor.kt
@@ -1,0 +1,13 @@
+package pl.allegro.tech.servicemesh.envoycontrol.snapshot
+
+import reactor.core.publisher.Mono
+
+interface SnapshotChangeAuditor {
+    fun audit(previousUpdateResult: UpdateResult, actualUpdateResult: UpdateResult): Mono<Any>
+}
+
+object NoopSnapshotChangeAuditor: SnapshotChangeAuditor {
+    override fun audit(previousUpdateResult: UpdateResult, actualUpdateResult: UpdateResult): Mono<Any> {
+        return Mono.empty()
+    }
+}