+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Lists all changes with user impact.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
 ### Changed
+- Global snapshot auditor added
 - updated project java version 11
 
 ## [0.19.5]